Update
Made the script more automated by automatically approving installs instead of additional user input (Created an Interactive Version for custom setups, i.e. you want to install RPM-Free but not RPM-NonFree, etc.)

Added the installation of the Flathub Repo to the script.

Added a comment to the beginning of the dnf.conf to show that the following code was added by this script (for future debugging purposes) Also added keepcache=True to the DNF Config to increase speed of DNF queries.